Migrated PLMath & PLGraphics & PLDatabase (+ backends) from c-style casts to C++ style casts (and again, fixed some const-cast messup, I really begin to like those C++ style casts :)

+++ PLBase/PLGraphics/src/Image/Loader/ImageLoaderPPM.cpp
@@ -83,7 +83,7 @@ bool ImageLoaderPPM::Load(Image &cImage, File &cFile)
 					// Loop through all integers (within error checks...)
 					uint8 *pImageData = pImageBuffer->GetData();
 					for (const uint8 *pImageDataEnd=pImageData+pImageBuffer->GetDataSize(); pImageData<pImageDataEnd; pImageData++)
-						*pImageData = (uint8)cTokenizer.GetNextToken().GetChar();
+						*pImageData = cTokenizer.GetNextToken().GetChar();
 
 					// Done
 					return true;
